set the page to scroll to the top when page reload and changed condition for scrollability on QA section
