+++ client/src/components/App.jsx
@@ -26,6 +26,7 @@ const App = function () {
 
   useEffect(() => {
     getAllData();
+    return () => { document.documentElement.scrollTop = 0 };
   }, [productId])
 
   if (isLoading) {
